requirements.txt
amqp==2.4.0
asgiref==2.3.2
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.1
channels==2.1.6
constantly==15.1.0
daphne==2.2.4
Django==2.1.5
django-extensions==2.1.4
django-ordered-model==3.1.1
djangorestframework==3.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
kombu==4.2.2.post1
PyHamcrest==1.9.0
pytz==2018.9
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
vine==1.2.0
zope.interface==4.6.0
